Abandoned scripts for localdb

+++ WorkoutTracker/Sql/Tables/DailyMetrics.sql
@@ -2,7 +2,7 @@ IF OBJECT_ID(N'Project.DailyMetrics') IS NULL
 BEGIN
     CREATE TABLE Project.DailyMetrics
     (   
-        MetricID FLOAT NOT NULL IDENTITY(1, 1) PRIMARY KEY,   
+        MetricID INT NOT NULL IDENTITY(1, 1) PRIMARY KEY,   
         Date NVARCHAR NOT NULL ,
         Weight FLOAT NOT NULL,     
         SleepDuration FLOAT NOT NULL,
